+++ example/lib/src/purchase_bloc/purchase_user_message.freezed.dart
@@ -1,0 +1,860 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'purchase_user_message.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
+
+/// @nodoc
+mixin _$PurchaseUserMessage {
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function() none,
+    required TResult Function() purchaseSuccess,
+    required TResult Function(ApphudError error) purchaseFailure,
+    required TResult Function() restorePurchasesSuccess,
+    required TResult Function(ApphudError error) restorePurchasesFailure,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function()? none,
+    TResult? Function()? purchaseSuccess,
+    TResult? Function(ApphudError error)? purchaseFailure,
+    TResult? Function()? restorePurchasesSuccess,
+    TResult? Function(ApphudError error)? restorePurchasesFailure,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? none,
+    TResult Function()? purchaseSuccess,
+    TResult Function(ApphudError error)? purchaseFailure,
+    TResult Function()? restorePurchasesSuccess,
+    TResult Function(ApphudError error)? restorePurchasesFailure,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(PurchaseNoneUserMessage value) none,
+    required TResult Function(PurchasePurchaseSuccessUserMessage value)
+        purchaseSuccess,
+    required TResult Function(PurchasePurchaseFailureUserMessage value)
+        purchaseFailure,
+    required TResult Function(PurchaseRestorePurchasesSuccessUserMessage value)
+        restorePurchasesSuccess,
+    required TResult Function(PurchaseRestorePurchasesFailureUserMessage value)
+        restorePurchasesFailure,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(PurchaseNoneUserMessage value)? none,
+    TResult? Function(PurchasePurchaseSuccessUserMessage value)?
+        purchaseSuccess,
+    TResult? Function(PurchasePurchaseFailureUserMessage value)?
+        purchaseFailure,
+    TResult? Function(PurchaseRestorePurchasesSuccessUserMessage value)?
+        restorePurchasesSuccess,
+    TResult? Function(PurchaseRestorePurchasesFailureUserMessage value)?
+        restorePurchasesFailure,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(PurchaseNoneUserMessage value)? none,
+    TResult Function(PurchasePurchaseSuccessUserMessage value)? purchaseSuccess,
+    TResult Function(PurchasePurchaseFailureUserMessage value)? purchaseFailure,
+    TResult Function(PurchaseRestorePurchasesSuccessUserMessage value)?
+        restorePurchasesSuccess,
+    TResult Function(PurchaseRestorePurchasesFailureUserMessage value)?
+        restorePurchasesFailure,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $PurchaseUserMessageCopyWith<$Res> {
+  factory $PurchaseUserMessageCopyWith(
+          PurchaseUserMessage value, $Res Function(PurchaseUserMessage) then) =
+      _$PurchaseUserMessageCopyWithImpl<$Res, PurchaseUserMessage>;
+}
+
+/// @nodoc
+class _$PurchaseUserMessageCopyWithImpl<$Res, $Val extends PurchaseUserMessage>
+    implements $PurchaseUserMessageCopyWith<$Res> {
+  _$PurchaseUserMessageCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+}
+
+/// @nodoc
+abstract class _$$PurchaseNoneUserMessageImplCopyWith<$Res> {
+  factory _$$PurchaseNoneUserMessageImplCopyWith(
+          _$PurchaseNoneUserMessageImpl value,
+          $Res Function(_$PurchaseNoneUserMessageImpl) then) =
+      __$$PurchaseNoneUserMessageImplCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class __$$PurchaseNoneUserMessageImplCopyWithImpl<$Res>
+    extends _$PurchaseUserMessageCopyWithImpl<$Res,
+        _$PurchaseNoneUserMessageImpl>
+    implements _$$PurchaseNoneUserMessageImplCopyWith<$Res> {
+  __$$PurchaseNoneUserMessageImplCopyWithImpl(
+      _$PurchaseNoneUserMessageImpl _value,
+      $Res Function(_$PurchaseNoneUserMessageImpl) _then)
+      : super(_value, _then);
+}
+
+/// @nodoc
+
+class _$PurchaseNoneUserMessageImpl extends PurchaseNoneUserMessage {
+  const _$PurchaseNoneUserMessageImpl() : super._();
+
+  @override
+  String toString() {
+    return 'PurchaseUserMessage.none()';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$PurchaseNoneUserMessageImpl);
+  }
+
+  @override
+  int get hashCode => runtimeType.hashCode;
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function() none,
+    required TResult Function() purchaseSuccess,
+    required TResult Function(ApphudError error) purchaseFailure,
+    required TResult Function() restorePurchasesSuccess,
+    required TResult Function(ApphudError error) restorePurchasesFailure,
+  }) {
+    return none();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function()? none,
+    TResult? Function()? purchaseSuccess,
+    TResult? Function(ApphudError error)? purchaseFailure,
+    TResult? Function()? restorePurchasesSuccess,
+    TResult? Function(ApphudError error)? restorePurchasesFailure,
+  }) {
+    return none?.call();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? none,
+    TResult Function()? purchaseSuccess,
+    TResult Function(ApphudError error)? purchaseFailure,
+    TResult Function()? restorePurchasesSuccess,
+    TResult Function(ApphudError error)? restorePurchasesFailure,
+    required TResult orElse(),
+  }) {
+    if (none != null) {
+      return none();
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(PurchaseNoneUserMessage value) none,
+    required TResult Function(PurchasePurchaseSuccessUserMessage value)
+        purchaseSuccess,
+    required TResult Function(PurchasePurchaseFailureUserMessage value)
+        purchaseFailure,
+    required TResult Function(PurchaseRestorePurchasesSuccessUserMessage value)
+        restorePurchasesSuccess,
+    required TResult Function(PurchaseRestorePurchasesFailureUserMessage value)
+        restorePurchasesFailure,
+  }) {
+    return none(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(PurchaseNoneUserMessage value)? none,
+    TResult? Function(PurchasePurchaseSuccessUserMessage value)?
+        purchaseSuccess,
+    TResult? Function(PurchasePurchaseFailureUserMessage value)?
+        purchaseFailure,
+    TResult? Function(PurchaseRestorePurchasesSuccessUserMessage value)?
+        restorePurchasesSuccess,
+    TResult? Function(PurchaseRestorePurchasesFailureUserMessage value)?
+        restorePurchasesFailure,
+  }) {
+    return none?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(PurchaseNoneUserMessage value)? none,
+    TResult Function(PurchasePurchaseSuccessUserMessage value)? purchaseSuccess,
+    TResult Function(PurchasePurchaseFailureUserMessage value)? purchaseFailure,
+    TResult Function(PurchaseRestorePurchasesSuccessUserMessage value)?
+        restorePurchasesSuccess,
+    TResult Function(PurchaseRestorePurchasesFailureUserMessage value)?
+        restorePurchasesFailure,
+    required TResult orElse(),
+  }) {
+    if (none != null) {
+      return none(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class PurchaseNoneUserMessage extends PurchaseUserMessage {
+  const factory PurchaseNoneUserMessage() = _$PurchaseNoneUserMessageImpl;
+  const PurchaseNoneUserMessage._() : super._();
+}
+
+/// @nodoc
+abstract class _$$PurchasePurchaseSuccessUserMessageImplCopyWith<$Res> {
+  factory _$$PurchasePurchaseSuccessUserMessageImplCopyWith(
+          _$PurchasePurchaseSuccessUserMessageImpl value,
+          $Res Function(_$PurchasePurchaseSuccessUserMessageImpl) then) =
+      __$$PurchasePurchaseSuccessUserMessageImplCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class __$$PurchasePurchaseSuccessUserMessageImplCopyWithImpl<$Res>
+    extends _$PurchaseUserMessageCopyWithImpl<$Res,
+        _$PurchasePurchaseSuccessUserMessageImpl>
+    implements _$$PurchasePurchaseSuccessUserMessageImplCopyWith<$Res> {
+  __$$PurchasePurchaseSuccessUserMessageImplCopyWithImpl(
+      _$PurchasePurchaseSuccessUserMessageImpl _value,
+      $Res Function(_$PurchasePurchaseSuccessUserMessageImpl) _then)
+      : super(_value, _then);
+}
+
+/// @nodoc
+
+class _$PurchasePurchaseSuccessUserMessageImpl
+    extends PurchasePurchaseSuccessUserMessage {
+  const _$PurchasePurchaseSuccessUserMessageImpl() : super._();
+
+  @override
+  String toString() {
+    return 'PurchaseUserMessage.purchaseSuccess()';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$PurchasePurchaseSuccessUserMessageImpl);
+  }
+
+  @override
+  int get hashCode => runtimeType.hashCode;
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function() none,
+    required TResult Function() purchaseSuccess,
+    required TResult Function(ApphudError error) purchaseFailure,
+    required TResult Function() restorePurchasesSuccess,
+    required TResult Function(ApphudError error) restorePurchasesFailure,
+  }) {
+    return purchaseSuccess();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function()? none,
+    TResult? Function()? purchaseSuccess,
+    TResult? Function(ApphudError error)? purchaseFailure,
+    TResult? Function()? restorePurchasesSuccess,
+    TResult? Function(ApphudError error)? restorePurchasesFailure,
+  }) {
+    return purchaseSuccess?.call();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? none,
+    TResult Function()? purchaseSuccess,
+    TResult Function(ApphudError error)? purchaseFailure,
+    TResult Function()? restorePurchasesSuccess,
+    TResult Function(ApphudError error)? restorePurchasesFailure,
+    required TResult orElse(),
+  }) {
+    if (purchaseSuccess != null) {
+      return purchaseSuccess();
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(PurchaseNoneUserMessage value) none,
+    required TResult Function(PurchasePurchaseSuccessUserMessage value)
+        purchaseSuccess,
+    required TResult Function(PurchasePurchaseFailureUserMessage value)
+        purchaseFailure,
+    required TResult Function(PurchaseRestorePurchasesSuccessUserMessage value)
+        restorePurchasesSuccess,
+    required TResult Function(PurchaseRestorePurchasesFailureUserMessage value)
+        restorePurchasesFailure,
+  }) {
+    return purchaseSuccess(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(PurchaseNoneUserMessage value)? none,
+    TResult? Function(PurchasePurchaseSuccessUserMessage value)?
+        purchaseSuccess,
+    TResult? Function(PurchasePurchaseFailureUserMessage value)?
+        purchaseFailure,
+    TResult? Function(PurchaseRestorePurchasesSuccessUserMessage value)?
+        restorePurchasesSuccess,
+    TResult? Function(PurchaseRestorePurchasesFailureUserMessage value)?
+        restorePurchasesFailure,
+  }) {
+    return purchaseSuccess?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(PurchaseNoneUserMessage value)? none,
+    TResult Function(PurchasePurchaseSuccessUserMessage value)? purchaseSuccess,
+    TResult Function(PurchasePurchaseFailureUserMessage value)? purchaseFailure,
+    TResult Function(PurchaseRestorePurchasesSuccessUserMessage value)?
+        restorePurchasesSuccess,
+    TResult Function(PurchaseRestorePurchasesFailureUserMessage value)?
+        restorePurchasesFailure,
+    required TResult orElse(),
+  }) {
+    if (purchaseSuccess != null) {
+      return purchaseSuccess(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class PurchasePurchaseSuccessUserMessage extends PurchaseUserMessage {
+  const factory PurchasePurchaseSuccessUserMessage() =
+      _$PurchasePurchaseSuccessUserMessageImpl;
+  const PurchasePurchaseSuccessUserMessage._() : super._();
+}
+
+/// @nodoc
+abstract class _$$PurchasePurchaseFailureUserMessageImplCopyWith<$Res> {
+  factory _$$PurchasePurchaseFailureUserMessageImplCopyWith(
+          _$PurchasePurchaseFailureUserMessageImpl value,
+          $Res Function(_$PurchasePurchaseFailureUserMessageImpl) then) =
+      __$$PurchasePurchaseFailureUserMessageImplCopyWithImpl<$Res>;
+  @useResult
+  $Res call({ApphudError error});
+}
+
+/// @nodoc
+class __$$PurchasePurchaseFailureUserMessageImplCopyWithImpl<$Res>
+    extends _$PurchaseUserMessageCopyWithImpl<$Res,
+        _$PurchasePurchaseFailureUserMessageImpl>
+    implements _$$PurchasePurchaseFailureUserMessageImplCopyWith<$Res> {
+  __$$PurchasePurchaseFailureUserMessageImplCopyWithImpl(
+      _$PurchasePurchaseFailureUserMessageImpl _value,
+      $Res Function(_$PurchasePurchaseFailureUserMessageImpl) _then)
+      : super(_value, _then);
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? error = null,
+  }) {
+    return _then(_$PurchasePurchaseFailureUserMessageImpl(
+      null == error
+          ? _value.error
+          : error // ignore: cast_nullable_to_non_nullable
+              as ApphudError,
+    ));
+  }
+}
+
+/// @nodoc
+
+class _$PurchasePurchaseFailureUserMessageImpl
+    extends PurchasePurchaseFailureUserMessage {
+  const _$PurchasePurchaseFailureUserMessageImpl(this.error) : super._();
+
+  @override
+  final ApphudError error;
+
+  @override
+  String toString() {
+    return 'PurchaseUserMessage.purchaseFailure(error: $error)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$PurchasePurchaseFailureUserMessageImpl &&
+            (identical(other.error, error) || other.error == error));
+  }
+
+  @override
+  int get hashCode => Object.hash(runtimeType, error);
+
+  @JsonKey(ignore: true)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$PurchasePurchaseFailureUserMessageImplCopyWith<
+          _$PurchasePurchaseFailureUserMessageImpl>
+      get copyWith => __$$PurchasePurchaseFailureUserMessageImplCopyWithImpl<
+          _$PurchasePurchaseFailureUserMessageImpl>(this, _$identity);
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function() none,
+    required TResult Function() purchaseSuccess,
+    required TResult Function(ApphudError error) purchaseFailure,
+    required TResult Function() restorePurchasesSuccess,
+    required TResult Function(ApphudError error) restorePurchasesFailure,
+  }) {
+    return purchaseFailure(error);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function()? none,
+    TResult? Function()? purchaseSuccess,
+    TResult? Function(ApphudError error)? purchaseFailure,
+    TResult? Function()? restorePurchasesSuccess,
+    TResult? Function(ApphudError error)? restorePurchasesFailure,
+  }) {
+    return purchaseFailure?.call(error);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? none,
+    TResult Function()? purchaseSuccess,
+    TResult Function(ApphudError error)? purchaseFailure,
+    TResult Function()? restorePurchasesSuccess,
+    TResult Function(ApphudError error)? restorePurchasesFailure,
+    required TResult orElse(),
+  }) {
+    if (purchaseFailure != null) {
+      return purchaseFailure(error);
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(PurchaseNoneUserMessage value) none,
+    required TResult Function(PurchasePurchaseSuccessUserMessage value)
+        purchaseSuccess,
+    required TResult Function(PurchasePurchaseFailureUserMessage value)
+        purchaseFailure,
+    required TResult Function(PurchaseRestorePurchasesSuccessUserMessage value)
+        restorePurchasesSuccess,
+    required TResult Function(PurchaseRestorePurchasesFailureUserMessage value)
+        restorePurchasesFailure,
+  }) {
+    return purchaseFailure(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(PurchaseNoneUserMessage value)? none,
+    TResult? Function(PurchasePurchaseSuccessUserMessage value)?
+        purchaseSuccess,
+    TResult? Function(PurchasePurchaseFailureUserMessage value)?
+        purchaseFailure,
+    TResult? Function(PurchaseRestorePurchasesSuccessUserMessage value)?
+        restorePurchasesSuccess,
+    TResult? Function(PurchaseRestorePurchasesFailureUserMessage value)?
+        restorePurchasesFailure,
+  }) {
+    return purchaseFailure?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(PurchaseNoneUserMessage value)? none,
+    TResult Function(PurchasePurchaseSuccessUserMessage value)? purchaseSuccess,
+    TResult Function(PurchasePurchaseFailureUserMessage value)? purchaseFailure,
+    TResult Function(PurchaseRestorePurchasesSuccessUserMessage value)?
+        restorePurchasesSuccess,
+    TResult Function(PurchaseRestorePurchasesFailureUserMessage value)?
+        restorePurchasesFailure,
+    required TResult orElse(),
+  }) {
+    if (purchaseFailure != null) {
+      return purchaseFailure(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class PurchasePurchaseFailureUserMessage extends PurchaseUserMessage {
+  const factory PurchasePurchaseFailureUserMessage(final ApphudError error) =
+      _$PurchasePurchaseFailureUserMessageImpl;
+  const PurchasePurchaseFailureUserMessage._() : super._();
+
+  ApphudError get error;
+  @JsonKey(ignore: true)
+  _$$PurchasePurchaseFailureUserMessageImplCopyWith<
+          _$PurchasePurchaseFailureUserMessageImpl>
+      get copyWith => throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class _$$PurchaseRestorePurchasesSuccessUserMessageImplCopyWith<$Res> {
+  factory _$$PurchaseRestorePurchasesSuccessUserMessageImplCopyWith(
+          _$PurchaseRestorePurchasesSuccessUserMessageImpl value,
+          $Res Function(_$PurchaseRestorePurchasesSuccessUserMessageImpl)
+              then) =
+      __$$PurchaseRestorePurchasesSuccessUserMessageImplCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class __$$PurchaseRestorePurchasesSuccessUserMessageImplCopyWithImpl<$Res>
+    extends _$PurchaseUserMessageCopyWithImpl<$Res,
+        _$PurchaseRestorePurchasesSuccessUserMessageImpl>
+    implements _$$PurchaseRestorePurchasesSuccessUserMessageImplCopyWith<$Res> {
+  __$$PurchaseRestorePurchasesSuccessUserMessageImplCopyWithImpl(
+      _$PurchaseRestorePurchasesSuccessUserMessageImpl _value,
+      $Res Function(_$PurchaseRestorePurchasesSuccessUserMessageImpl) _then)
+      : super(_value, _then);
+}
+
+/// @nodoc
+
+class _$PurchaseRestorePurchasesSuccessUserMessageImpl
+    extends PurchaseRestorePurchasesSuccessUserMessage {
+  const _$PurchaseRestorePurchasesSuccessUserMessageImpl() : super._();
+
+  @override
+  String toString() {
+    return 'PurchaseUserMessage.restorePurchasesSuccess()';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$PurchaseRestorePurchasesSuccessUserMessageImpl);
+  }
+
+  @override
+  int get hashCode => runtimeType.hashCode;
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function() none,
+    required TResult Function() purchaseSuccess,
+    required TResult Function(ApphudError error) purchaseFailure,
+    required TResult Function() restorePurchasesSuccess,
+    required TResult Function(ApphudError error) restorePurchasesFailure,
+  }) {
+    return restorePurchasesSuccess();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function()? none,
+    TResult? Function()? purchaseSuccess,
+    TResult? Function(ApphudError error)? purchaseFailure,
+    TResult? Function()? restorePurchasesSuccess,
+    TResult? Function(ApphudError error)? restorePurchasesFailure,
+  }) {
+    return restorePurchasesSuccess?.call();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? none,
+    TResult Function()? purchaseSuccess,
+    TResult Function(ApphudError error)? purchaseFailure,
+    TResult Function()? restorePurchasesSuccess,
+    TResult Function(ApphudError error)? restorePurchasesFailure,
+    required TResult orElse(),
+  }) {
+    if (restorePurchasesSuccess != null) {
+      return restorePurchasesSuccess();
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(PurchaseNoneUserMessage value) none,
+    required TResult Function(PurchasePurchaseSuccessUserMessage value)
+        purchaseSuccess,
+    required TResult Function(PurchasePurchaseFailureUserMessage value)
+        purchaseFailure,
+    required TResult Function(PurchaseRestorePurchasesSuccessUserMessage value)
+        restorePurchasesSuccess,
+    required TResult Function(PurchaseRestorePurchasesFailureUserMessage value)
+        restorePurchasesFailure,
+  }) {
+    return restorePurchasesSuccess(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(PurchaseNoneUserMessage value)? none,
+    TResult? Function(PurchasePurchaseSuccessUserMessage value)?
+        purchaseSuccess,
+    TResult? Function(PurchasePurchaseFailureUserMessage value)?
+        purchaseFailure,
+    TResult? Function(PurchaseRestorePurchasesSuccessUserMessage value)?
+        restorePurchasesSuccess,
+    TResult? Function(PurchaseRestorePurchasesFailureUserMessage value)?
+        restorePurchasesFailure,
+  }) {
+    return restorePurchasesSuccess?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(PurchaseNoneUserMessage value)? none,
+    TResult Function(PurchasePurchaseSuccessUserMessage value)? purchaseSuccess,
+    TResult Function(PurchasePurchaseFailureUserMessage value)? purchaseFailure,
+    TResult Function(PurchaseRestorePurchasesSuccessUserMessage value)?
+        restorePurchasesSuccess,
+    TResult Function(PurchaseRestorePurchasesFailureUserMessage value)?
+        restorePurchasesFailure,
+    required TResult orElse(),
+  }) {
+    if (restorePurchasesSuccess != null) {
+      return restorePurchasesSuccess(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class PurchaseRestorePurchasesSuccessUserMessage
+    extends PurchaseUserMessage {
+  const factory PurchaseRestorePurchasesSuccessUserMessage() =
+      _$PurchaseRestorePurchasesSuccessUserMessageImpl;
+  const PurchaseRestorePurchasesSuccessUserMessage._() : super._();
+}
+
+/// @nodoc
+abstract class _$$PurchaseRestorePurchasesFailureUserMessageImplCopyWith<$Res> {
+  factory _$$PurchaseRestorePurchasesFailureUserMessageImplCopyWith(
+          _$PurchaseRestorePurchasesFailureUserMessageImpl value,
+          $Res Function(_$PurchaseRestorePurchasesFailureUserMessageImpl)
+              then) =
+      __$$PurchaseRestorePurchasesFailureUserMessageImplCopyWithImpl<$Res>;
+  @useResult
+  $Res call({ApphudError error});
+}
+
+/// @nodoc
+class __$$PurchaseRestorePurchasesFailureUserMessageImplCopyWithImpl<$Res>
+    extends _$PurchaseUserMessageCopyWithImpl<$Res,
+        _$PurchaseRestorePurchasesFailureUserMessageImpl>
+    implements _$$PurchaseRestorePurchasesFailureUserMessageImplCopyWith<$Res> {
+  __$$PurchaseRestorePurchasesFailureUserMessageImplCopyWithImpl(
+      _$PurchaseRestorePurchasesFailureUserMessageImpl _value,
+      $Res Function(_$PurchaseRestorePurchasesFailureUserMessageImpl) _then)
+      : super(_value, _then);
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? error = null,
+  }) {
+    return _then(_$PurchaseRestorePurchasesFailureUserMessageImpl(
+      null == error
+          ? _value.error
+          : error // ignore: cast_nullable_to_non_nullable
+              as ApphudError,
+    ));
+  }
+}
+
+/// @nodoc
+
+class _$PurchaseRestorePurchasesFailureUserMessageImpl
+    extends PurchaseRestorePurchasesFailureUserMessage {
+  const _$PurchaseRestorePurchasesFailureUserMessageImpl(this.error)
+      : super._();
+
+  @override
+  final ApphudError error;
+
+  @override
+  String toString() {
+    return 'PurchaseUserMessage.restorePurchasesFailure(error: $error)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$PurchaseRestorePurchasesFailureUserMessageImpl &&
+            (identical(other.error, error) || other.error == error));
+  }
+
+  @override
+  int get hashCode => Object.hash(runtimeType, error);
+
+  @JsonKey(ignore: true)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$PurchaseRestorePurchasesFailureUserMessageImplCopyWith<
+          _$PurchaseRestorePurchasesFailureUserMessageImpl>
+      get copyWith =>
+          __$$PurchaseRestorePurchasesFailureUserMessageImplCopyWithImpl<
+                  _$PurchaseRestorePurchasesFailureUserMessageImpl>(
+              this, _$identity);
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function() none,
+    required TResult Function() purchaseSuccess,
+    required TResult Function(ApphudError error) purchaseFailure,
+    required TResult Function() restorePurchasesSuccess,
+    required TResult Function(ApphudError error) restorePurchasesFailure,
+  }) {
+    return restorePurchasesFailure(error);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function()? none,
+    TResult? Function()? purchaseSuccess,
+    TResult? Function(ApphudError error)? purchaseFailure,
+    TResult? Function()? restorePurchasesSuccess,
+    TResult? Function(ApphudError error)? restorePurchasesFailure,
+  }) {
+    return restorePurchasesFailure?.call(error);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? none,
+    TResult Function()? purchaseSuccess,
+    TResult Function(ApphudError error)? purchaseFailure,
+    TResult Function()? restorePurchasesSuccess,
+    TResult Function(ApphudError error)? restorePurchasesFailure,
+    required TResult orElse(),
+  }) {
+    if (restorePurchasesFailure != null) {
+      return restorePurchasesFailure(error);
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(PurchaseNoneUserMessage value) none,
+    required TResult Function(PurchasePurchaseSuccessUserMessage value)
+        purchaseSuccess,
+    required TResult Function(PurchasePurchaseFailureUserMessage value)
+        purchaseFailure,
+    required TResult Function(PurchaseRestorePurchasesSuccessUserMessage value)
+        restorePurchasesSuccess,
+    required TResult Function(PurchaseRestorePurchasesFailureUserMessage value)
+        restorePurchasesFailure,
+  }) {
+    return restorePurchasesFailure(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(PurchaseNoneUserMessage value)? none,
+    TResult? Function(PurchasePurchaseSuccessUserMessage value)?
+        purchaseSuccess,
+    TResult? Function(PurchasePurchaseFailureUserMessage value)?
+        purchaseFailure,
+    TResult? Function(PurchaseRestorePurchasesSuccessUserMessage value)?
+        restorePurchasesSuccess,
+    TResult? Function(PurchaseRestorePurchasesFailureUserMessage value)?
+        restorePurchasesFailure,
+  }) {
+    return restorePurchasesFailure?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(PurchaseNoneUserMessage value)? none,
+    TResult Function(PurchasePurchaseSuccessUserMessage value)? purchaseSuccess,
+    TResult Function(PurchasePurchaseFailureUserMessage value)? purchaseFailure,
+    TResult Function(PurchaseRestorePurchasesSuccessUserMessage value)?
+        restorePurchasesSuccess,
+    TResult Function(PurchaseRestorePurchasesFailureUserMessage value)?
+        restorePurchasesFailure,
+    required TResult orElse(),
+  }) {
+    if (restorePurchasesFailure != null) {
+      return restorePurchasesFailure(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class PurchaseRestorePurchasesFailureUserMessage
+    extends PurchaseUserMessage {
+  const factory PurchaseRestorePurchasesFailureUserMessage(
+          final ApphudError error) =
+      _$PurchaseRestorePurchasesFailureUserMessageImpl;
+  const PurchaseRestorePurchasesFailureUserMessage._() : super._();
+
+  ApphudError get error;
+  @JsonKey(ignore: true)
+  _$$PurchaseRestorePurchasesFailureUserMessageImplCopyWith<
+          _$PurchaseRestorePurchasesFailureUserMessageImpl>
+      get copyWith => throw _privateConstructorUsedError;
+}