+++ example/lib/src/view/widgets/paywalls_list_widget.dart
@@ -1,7 +1,6 @@
 import 'package:apphud/models/apphud_models/apphud_paywall.dart';
-import 'package:apphud_example/src/feature/home/product_list_widget.dart';
-import 'package:apphud_example/src/feature/purchase/purchase_bloc.dart';
-import 'package:apphud_example/src/feature/purchase/purchase_event.dart';
+import 'package:apphud_example/src/purchase_bloc/purchase_bloc.dart';
+import 'package:apphud_example/src/view/widgets/product_list_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -85,7 +84,7 @@ class _PaywallsListWidgetState extends State<PaywallsListWidget> {
       key: ValueKey(paywall.identifier),
       title: Text(
         paywall.identifier,
-        style: Theme.of(context).textTheme.headline5,
+        style: Theme.of(context).textTheme.headlineSmall,
       ),
       subtitle: Text(
         'products: ${paywall.products?.length ?? 0}\n'